backend/requirements.txt
# =============================================================================
# Zeroback Fashion AI Backend Dependencies
# 최적화된 의존성 목록 (2025.08.19)
# =============================================================================

# Web Framework & API
Flask==3.0.3
Flask-CORS==4.0.1
flasgger==0.9.7.1
Werkzeug==3.0.3

# Data Processing & Utilities
PyYAML==6.0.1
numpy>=1.24.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0

# AI/ML Libraries
torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<1.0.0
ultralytics>=8.0.0,<9.0.0

# Image Processing
opencv-python-headless>=4.8.0,<5.0.0
Pillow>=10.0.0,<11.0.0

# External API Services
google-generativeai>=0.7.0,<1.0.0
requests>=2.31.0,<3.0.0

# Development & Testing (optional)
pytest>=7.0.0,<8.0.0
pytest-flask>=1.0.0,<2.0.0